node/meta: move block related things in a separate file
Also rename `container.go` to `containers.go` since we already have `blocks.go`
and `notifications.go` with the same meaning.
